requirements.txt
python-dotenv
openai
pydantic
anthropic
groq
google-generativeai
matplotlib
pandas
seaborn
ipykernel